checked shell scripts with shellcheck and python scripts with black
