+++ rook/kubestone_fio/scripts/run_benchmarks.sh
@@ -2,7 +2,7 @@
 set -eu
 
 function info {
-  echo "[info] - $(date) - $@"
+  echo "[info] - $(date +%r) - $*"
 }
 
 export PYTHONIOENCODING=UTF-8
@@ -15,16 +15,16 @@ sleep 5
 done
 
 targets=""
-for blkdev in ${BLOCKDEVICES[@]};
+for blkdev in "${BLOCKDEVICES[@]}";
 do
 info "formatting /dev/${blkdev}"
-kubectl exec -n kubestone ${pod} -- mkfs.ext4 /dev/${blkdev}
+kubectl exec -n kubestone "$pod" -- mkfs.ext4 "/dev/${blkdev}"
 targets+=" /dev/${blkdev} "
 done
 
 info "dropping caches"
-kubectl exec -n kubestone ${pod} -- sh -c "echo 3 | tee /proc/sys/vm/drop_caches"
+kubectl exec -n kubestone "$pod" -- sh -c "echo 3 | tee /proc/sys/vm/drop_caches"
 bench_fio_cmd="bench_fio --target ${targets} --template /fio-plot/benchmark_script/fio-job-template.fio --type device --mode ${MODES} --output FIO_OUTPUT -e ${IO_ENGINE} -b ${BLOCKSIZE} --iodepth ${IO_DEPTH} --numjobs ${NUM_JOBS}"
 
 info "running benchmark"
-kubectl exec -n kubestone ${pod}  -- ${bench_fio_cmd}
+kubectl exec -n kubestone "$pod" -- sh -c "$bench_fio_cmd"
